[State Sync] Move fuzzer into self-contained file.
